remove dead code + tidy up settings
