+++ frontend/src/metabase/admin/people/components/PeopleNav.tsx
@@ -1,11 +1,11 @@
 import { t } from "ttag";
 
-import { shouldNudgeToPro } from "metabase/admin/people/selectors";
 import {
   AdminNavItem,
   type AdminNavItemProps,
   AdminNavWrapper,
-} from "metabase/admin/settings/components/AdminNav";
+} from "metabase/admin/AdminNav";
+import { shouldNudgeToPro } from "metabase/admin/people/selectors";
 import { UpsellSSO } from "metabase/admin/upsells";
 import { useSelector } from "metabase/lib/redux";
 import { getLocation } from "metabase/selectors/routing";
