+++ frontend/src/metabase/admin/performance/components/PerformanceApp.tsx
@@ -1,7 +1,7 @@
 import {
   AdminNavItem,
   AdminNavWrapper,
-} from "metabase/admin/settings/components/AdminNav";
+} from "metabase/admin/AdminNav";
 import { AdminSettingsLayout } from "metabase/common/components/AdminLayout/AdminSettingsLayout";
 import { PLUGIN_CACHING } from "metabase/plugins";
