--- frontend/src/metabase/admin/settings/app/components/SettingsEditor/index.ts
@@ -1,1 +1,0 @@
-export { default as SettingsEditor } from "./SettingsEditor";
