+++ frontend/src/metabase/admin/settings/components/widgets/GroupMappingsWidget/GroupMappingsWidget.tsx
@@ -1,12 +1,13 @@
 import _ from "underscore";
 
-import GroupMappingsWidget from "metabase/admin/settings/components/widgets/GroupMappingsWidget/GroupMappingsWidget";
 import { updateSetting } from "metabase/admin/settings/settings";
 import Group from "metabase/entities/groups";
 import { connect } from "metabase/lib/redux";
 import { getSetting } from "metabase/selectors/settings";
 import type { Settings } from "metabase-types/api/settings";
 import type { State } from "metabase-types/store";
+
+import { GroupMappingsWidgetView } from "./GroupMappingsWidgetView";
 
 const mapStateToProps = (
   state: State,
@@ -24,8 +25,7 @@ const mapDispatchToProps = {
   clearGroupMember: Group.actions.clearMember,
 };
 
-// eslint-disable-next-line import/no-default-export -- deprecated usage
-export default _.compose(
+export const GroupMappingsWidget = _.compose(
   connect(mapStateToProps, mapDispatchToProps),
   Group.loadList(),
-)(GroupMappingsWidget);
+)(GroupMappingsWidgetView);
