+++ frontend/src/metabase/admin/settings/components/widgets/GroupMappingsWidget/index.ts
@@ -1,0 +1,1 @@
+export * from "./GroupMappingsWidget";
