--- frontend/src/metabase/admin/settings/components/widgets/SettingInput/SettingInput.styled.tsx
@@ -1,8 +1,0 @@
-// eslint-disable-next-line no-restricted-imports
-import styled from "@emotion/styled";
-
-import InputBlurChange from "metabase/common/components/InputBlurChange";
-
-export const SettingInputBlurChange = styled(InputBlurChange)`
-  width: 400px;
-`;
