--- frontend/src/metabase/admin/settings/components/widgets/SettingInput/index.ts
@@ -1,1 +1,0 @@
-export { SettingInput } from "./SettingInput";
