--- frontend/src/metabase/admin/settings/components/widgets/SettingNumber.jsx
@@ -1,8 +1,0 @@
-/* eslint-disable react/prop-types */
-import { SettingInput } from "./SettingInput";
-
-const SettingNumber = ({ type = "number", ...props }) => (
-  <SettingInput {...props} type="number" />
-);
-
-export default SettingNumber;
