--- frontend/src/metabase/admin/settings/components/widgets/SettingPassword.jsx
@@ -1,5 +1,0 @@
-import { SettingInput } from "./SettingInput";
-
-const SettingPassword = (props) => <SettingInput {...props} type="password" />;
-
-export default SettingPassword;
