--- frontend/src/metabase/admin/settings/components/widgets/SettingTextInput/index.ts
@@ -1,1 +1,0 @@
-export { SettingTextInput } from "./SettingTextInput";
