+++ frontend/src/metabase/admin/settings/components/widgets/VersionUpdateNotice/VersionUpdateNotice.tsx
@@ -1,7 +1,7 @@
 import cx from "classnames";
 import { c, t } from "ttag";
 
-import { getCurrentVersion } from "metabase/admin/settings/selectors";
+import { getCurrentVersion } from "metabase/admin/app/selectors";
 import { useGetVersionInfoQuery } from "metabase/api";
 import ExternalLink from "metabase/common/components/ExternalLink";
 import { useSetting } from "metabase/common/hooks";
