--- frontend/src/metabase/admin/settings/selectors/index.ts
@@ -1,2 +1,0 @@
-export * from "./selectors";
-export * from "./typed-selectors";
