--- frontend/src/metabase/admin/settings/selectors/typed-selectors.ts
@@ -1,4 +1,0 @@
-import type { State } from "metabase-types/store";
-
-export const getAdminSettingDefinitions = (state: State) =>
-  state.admin.settings.settings;
